+++ src/Service.php
@@ -20,6 +20,8 @@ abstract class Service
   use HasRequestHooks;
   use HasApiSearch;
   use HasPluginList;
+  use HasAuthentication;
+  use HasAuthorization;
   use HasWire;
   use HasLock;