+++ src/args/EndpointHandlerArgs.php
@@ -12,6 +12,9 @@ class EndpointHandlerArgs
   /** Request */
   public Request $request;
 
+  /** ProcessWire user */
+  public \ProcessWire\User $user;
+
   /** ProcessWire URL hook event */
   public \ProcessWire\HookEvent $event;
 }