+++ src/args/ExceptionHandlerArgs.php
@@ -15,6 +15,9 @@ class ExceptionHandlerArgs
   /** Request */
   public Request $request;
 
+  /** ProcessWire user */
+  public \ProcessWire\User $user;
+
   /** ProcessWire URL hook event */
   public \ProcessWire\HookEvent $event;